fix(skit-authoring): 執筆ツールの原点解決をSkitManager経由にしInspectorの途中returnをやめる
原点をClientDIContext限定で引いていたためSkitTest.unityでは常に解決に失敗し、
座標コピーが恒久的に拒否されていた。シーン上のSkitManagerから引くようにして
本編とSkitTestを同一経路に揃えた。

あわせてボタン押下時の途中returnをやめ、原点が解決できない場合でも
base.OnInspectorGUI()へ必ず到達するようにした（IMGUIのレイアウト不整合を防ぐ）。
Editor/Inspectorが12ファイルになっていたためSkit執筆系3ファイルを
Editor/Inspector/Skit/へ移した。
