+++ moorestech_client/Assets/Scripts/Editor/Inspector/Skit.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d6228f3c6c924050b22e5c7d63a37a9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
