+++ moorestech_client/Assets/Scripts/Editor/Inspector/Skit/SkitAuthoringOriginResolver.cs.meta
