+++ moorestech_client/Assets/Scripts/Editor/Inspector/Skit/SkitCharacterEditorUtilInspector.cs
@@ -3,34 +3,34 @@ using Client.Skit.Skit;
 using UnityEditor;
 using UnityEngine;
 
-[CustomEditor(typeof(SkitCameraEditorUtil))]
-public class SkitCameraEditorUtilInspector : Editor
+[CustomEditor(typeof(SkitCharacterEditorUtil))]
+public class SkitCharacterEditorUtilInspector : Editor
 {
     public override void OnInspectorGUI()
     {
-        var obj = (SkitCameraEditorUtil)target;
+        var obj = (SkitCharacterEditorUtil)target;
         
-        if (GUILayout.Button("カメラワープコマンドをコピー"))
+        if (GUILayout.Button("キャラ位置設定コマンドをコピー"))
         {
             // スキットJSONは相対座標なのでPlayMode中の実原点で変換する（ADR 0029）
             // Skit JSON is spawn-relative, so convert with the live origin during PlayMode (ADR 0029)
-            if (!SkitAuthoringOriginResolver.TryResolve(out var origin))
-            {
-                return;
-            }
-            
+            if (SkitAuthoringOriginResolver.TryResolve(out var origin)) CopyCommand(origin);
+        }
+        
+        base.OnInspectorGUI();
+        
+        #region Internal
+        
+        void CopyCommand(SkitOrigin origin)
+        {
             var pos = origin.ToRelative(obj.transform.position);
             var rot = obj.transform.eulerAngles;
             
-            // Camera コンポーネントを取得
-            var cam = obj.GetComponent<Camera>();
-            var fov = cam != null ? cam.fieldOfView : 0f;
-            
-            string str = $@"[
+            var str = $@"[
     {{
-        ""type"": ""cameraWarp"",
+        ""type"": ""characterTransform"",
         ""backgroundColor"": ""#ffffff"",
-        ""fieldOfView"": {fov},
+        ""character"": ""{obj.characterId}"",
         ""Position"": [
             {pos.x},
             {pos.y},
@@ -44,9 +44,9 @@ public class SkitCameraEditorUtilInspector : Editor
         ""id"": 1
     }}
 ]";
-            EditorGUIUtility.systemCopyBuffer = str;  // クリップボードへコピー
+            EditorGUIUtility.systemCopyBuffer = str;   // クリップボードへコピー
         }
         
-        base.OnInspectorGUI();
+        #endregion
     }
 }
